app/vectorstore.py
import os
from dotenv import load_dotenv
from langchain_openai import OpenAIEmbeddings
from langchain_community.vectorstores import Chroma

load_dotenv()

def get_vectorstore():
    persist_directory = os.getenv("CHROMA_PERSIST_DIRECTORY", "./chroma_db")
    embeddings = OpenAIEmbeddings()
    return Chroma(
        persist_directory=persist_directory,
        embedding_function=embeddings
    )
