requirements.txt
O365
langchain
langchain-community
langchain-openai
chromadb
openai
fastapi
uvicorn
apscheduler
python-dotenv
python-docx
PyMuPDF
